+++ agendando_um_jeito/src/contexts/authContext.jsx
@@ -1,12 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import { onAuthStateChanged, signOut } from 'firebase/auth';
-import { auth } from '../services/firebase';
+import { createUserWithEmailAndPassword } from "firebase/auth";
+import { auth, db } from "../services/firebase";
+import { doc, setDoc } from "firebase/firestore";
 
 export const AuthContext = React.createContext({});
 
 export const AuthProvider = ({ children }) => {
     const [signed, setSigned] = useState(false)
     const [user, setUser] = useState(null);
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
 
     useEffect(() => {
         const unsubscribe = onAuthStateChanged(auth, (user) => {
@@ -17,8 +21,6 @@ export const AuthProvider = ({ children }) => {
                 setUser(null);
             }
         })
-
-        // * cleanup;
         return unsubscribe;
     }, [])
 
@@ -27,7 +29,24 @@ export const AuthProvider = ({ children }) => {
         return signOut(auth)
     }
 
-    const value = { signed, setSigned, user, setUser, logout };
+    const Signup = async (e) => {
+        e.preventDefault();
+
+        try {
+            const userCredential = await createUserWithEmailAndPassword(auth, email, password)
+            const user = userCredential.user;
+
+            const usersCollectionRef = doc(db, 'users', user.uid);
+            await setDoc(usersCollectionRef, { email, password })
+
+            setEmail("");
+            setPassword("");
+        } catch (error) {
+            console.log('error: ', error);
+        }
+    }
+
+    const value = { signed, setSigned, user, setUser, logout, Signup, email, setEmail, password, setPassword };
 
     return (
         <AuthContext.Provider value={value}>
